Second version.
Not sure how to push this into the previous version, since I didn't pull the previous one before I made changes. I guess I need to find a github tutorial.
Anyhow, this is the version I was talking to you about, Mike. There are notes in the code about the issue I'm having. I may be beating a dead horse in trying to fix this without the obvious way to rewrite it, but I'm curious to see if there's a way to do it. If you don't have time to screw around with it, I understand.